# swagger.yaml
openapi: 3.0.0
info:
  title: Workout Management API
  description: A complete RESTful API backend for a gym workout tracking application built with Node.js, Express, and PostgreSQL.
  version: 1.0.0
  contact:
    email: support@workoutapi.com
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_user:
          type: integer
          example: 1
        name:
          type: string
          example: "Bench Press"
        description:
          type: string
          nullable: true
          example: "Chest exercise"
        image_url:
          type: string
          nullable: true
          example: "/images/exercises/bench-press.jpg"
        has_pulley:
          type: boolean
          default: false
        is_unilateral:
          type: boolean
          default: false
        bar_weight:
          type: integer
          nullable: true
          enum: [6, 8, 10, 12, 20]
          example: 20
        muscle_group:
          type: string
          example: "Chest"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Workout:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_user:
          type: integer
          example: 1
        name:
          type: string
          example: "Push Day"
        description:
          type: string
          nullable: true
          example: "Chest, shoulders and triceps"
        image_url:
          type: string
          nullable: true
          example: "/images/workouts/push-day.jpg"
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutExercise'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkoutExercise:
      type: object
      properties:
        id:
          type: integer
        id_exercise:
          type: integer
        sets:
          type: integer
          example: 3
        number:
          type: integer
          example: 1
        name:
          type: string
          example: "Bench Press"
        muscle_group:
          type: string
          example: "Chest"
    Routine:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_user:
          type: integer
          example: 1
        name:
          type: string
          example: "PPL Split"
        description:
          type: string
          nullable: true
          example: "Push Pull Legs routine"
        image_url:
          type: string
          nullable: true
          example: "/images/routines/ppl.jpg"
        workouts:
          type: array
          items:
            $ref: '#/components/schemas/RoutineWorkout'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RoutineWorkout:
      type: object
      properties:
        id:
          type: integer
        id_workout:
          type: integer
        number:
          type: integer
          example: 1
        name:
          type: string
          example: "Push Day"
        description:
          type: string
          nullable: true
    WorkoutSession:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_user:
          type: integer
          example: 1
        id_workout:
          type: integer
          example: 1
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
          nullable: true
        total_volume:
          type: integer
          example: 5000
        notes:
          type: string
          nullable: true
          example: "Good workout today"
        workout_name:
          type: string
          example: "Push Day"
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/SessionExercise'
        sets:
          type: array
          items:
            $ref: '#/components/schemas/SessionSet'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SessionExercise:
      type: object
      properties:
        id:
          type: integer
        id_exercise:
          type: integer
        number:
          type: integer
          example: 1
        observations:
          type: string
          nullable: true
          example: "Felt heavy today"
        name:
          type: string
          example: "Bench Press"
        muscle_group:
          type: string
          example: "Chest"
    SessionSet:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_workout_session:
          type: integer
          example: 1
        id_exercise:
          type: integer
          example: 1
        number_of_set:
          type: integer
          example: 1
        sets:
          type: integer
          example: 1
        series_type:
          type: integer
          enum: [0, 1, 2, 3]
          example: 2
          description: "0: Warm-up, 1: Preparatory, 2: Working set, 3: Back off set"
        weight:
          type: integer
          example: 80
        reps:
          type: integer
          example: 10
        done:
          type: boolean
          default: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /api/exercises:
    get:
      tags:
        - Exercises
      summary: Get all user exercises
      security:
        - BearerAuth: []
      parameters:
        - name: muscle_group
          in: query
          description: Filter by muscle group
          schema:
            type: string
            example: "Chest"
      responses:
        '200':
          description: Exercises retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  exercises:
                    type: array
                    items:
                      $ref: '#/components/schemas/Exercise'
    post:
      tags:
        - Exercises
      summary: Create new exercise
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - muscle_group
              properties:
                name:
                  type: string
                  example: "Bench Press"
                description:
                  type: string
                  example: "Chest exercise"
                muscle_group:
                  type: string
                  example: "Chest"
                has_pulley:
                  type: boolean
                  default: false
                is_unilateral:
                  type: boolean
                  default: false
                bar_weight:
                  type: integer
                  enum: [6, 8, 10, 12, 20]
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Exercise created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Exercise created successfully"
                  exercise:
                    $ref: '#/components/schemas/Exercise'
...
  /api/workouts:
    get:
      tags:
        - Workouts
      summary: Get all user workouts
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Workouts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  workouts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workout'
    post:
      tags:
        - Workouts
      summary: Create new workout
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "Push Day"
                description:
                  type: string
                  example: "Chest, shoulders and triceps"
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Workout created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Workout created successfully"
                  workout:
                    $ref: '#/components/schemas/Workout'
...
  /api/routines:
    get:
      tags:
        - Routines
      summary: Get all user routines
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Routines retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  routines:
                    type: array
                    items:
                      $ref: '#/components/schemas/Routine'
    post:
      tags:
        - Routines
      summary: Create new routine
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "PPL Split"
                description:
                  type: string
                  example: "Push Pull Legs routine"
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Routine created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Routine created successfully"
                  routine:
                    $ref: '#/components/schemas/Routine'
...
  /api/workout-sessions:
    get:
      tags:
        - Workout Sessions
      summary: Get user workout sessions
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Number of sessions to return
          schema:
            type: integer
            default: 20
            example: 10
        - name: offset
          in: query
          description: Number of sessions to skip
          schema:
            type: integer
            default: 0
            example: 0
      responses:
        '200':
          description: Workout sessions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkoutSession'
    post:
      tags:
        - Workout Sessions
      summary: Start new workout session
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - workout_id
              properties:
                workout_id:
                  type: integer
                  example: 1
                notes:
                  type: string
                  example: "Morning workout"
      responses:
        '201':
          description: Workout session started successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Workout session started successfully"
                  session:
                    $ref: '#/components/schemas/WorkoutSession'
...